+++ src/Pages/Dashboard/components/EditProduct.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import ProductService from '../../services/ProductService';
+import ProductService from '../../../services/ProductService';
 import { MdDeleteForever } from 'react-icons/md';
 import { toast } from 'react-toastify';
