+++ src/Pages/Dashboard/components/NewService.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import ServiceService from '../../services/ServiceService';
+import ServiceService from '../../../services/ServiceService';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
