+++ src/Pages/Dashboard/components/NewStaff.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import StaffService from '../../services/StaffService';
+import StaffService from '../../../services/StaffService';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
