+++ src/Pages/Dashboard/components/Services.js
@@ -1,5 +1,4 @@
 import React from "react";
-import SubTitle from "../SubTitle"
 import ServicesTable from "./ServicesTable";
 import { FaPlus } from 'react-icons/fa';
 import { Link } from "react-router-dom";
