+++ src/Pages/Dashboard/components/ServicesTable.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../styles/Table.css";
+import "../../../styles/Table.css";
 import { Link } from "react-router-dom";
 import { FaEdit } from 'react-icons/fa';
