+++ src/Pages/Sales/components/ServiceInfo.js
@@ -1,5 +1,5 @@
 import React from "react";
-import SubTitle from '../SubTitle'
+import SubTitle from '../../../components/SubTitle'
 
 const ServiceInfo = ({ service }) => {
     return (
